backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.41
alembic==1.16.4
pydantic==2.11.7
dspy-ai==2.6.27
python-dotenv==1.1.1
python-multipart==0.0.20
httpx==0.28.1
pytest==8.4.1
pytest-asyncio==1.1.0
ruff==0.12.4
